refactor(starling): split monolith, fix CORS/PIN/styles, add net.fetch IPC
- Add generic sero.net.fetch IPC proxy (bypasses CORS for all apps)
- Refactor 1862-line StarlingApp.tsx into 14 files (all <500 LOC)
- Fix PIN pad layout: add missing styles.css with Tailwind + @theme inline
- Fix token persistence: Lock keeps encrypted token, only PIN needed on relaunch
- Add Forget Account flow with typed confirmation (FORGET MY ACCOUNT)
- Redesign color scheme: dark navy + emerald-green accent (matches weight-tracker)
- Fix dashboard header padding
- Add missing ui/index.html (fixes vite build entry resolution)
- Document index.html requirement in apps-tutorial.md
