+++ apps/desktop/electron/ipc/index.ts
@@ -28,6 +28,7 @@ import { registerVcsHandlers } from './vcs';
 import { registerGitHubHandlers } from './github';
 import { registerFeedbackHandlers } from './feedback';
 import { registerImagegenHandlers } from './imagegen';
+import { registerNetHandlers } from './net';
 import { registerUserFeedbackQuestionHandlers } from './user-feedback-questions';
 
 export function registerAllIpcHandlers(): void {
@@ -53,5 +54,6 @@ export function registerAllIpcHandlers(): void {
   registerGitHubHandlers();
   registerFeedbackHandlers();
   registerImagegenHandlers();
+  registerNetHandlers();
   registerUserFeedbackQuestionHandlers();
 }
