+++ apps/desktop/electron/preload.ts
@@ -25,6 +25,8 @@ import type {
   ResponseFeedbackEntry,
   ResponseFeedbackState,
   ChatAttachment,
+  ProxyFetchRequest,
+  ProxyFetchResponse,
 } from '../src/types/ipc';
 import type {
   VcsCheckpoint,
@@ -424,6 +426,11 @@ contextBridge.exposeInMainWorld('sero', {
       ipcRenderer.invoke(IpcChannels.layout.save, state),
     load: (): Promise<{ mainSidebarOpen: boolean; chatPanelOpen: boolean } | null> =>
       ipcRenderer.invoke(IpcChannels.layout.load),
+  },
+
+  net: {
+    fetch: (request: ProxyFetchRequest): Promise<ProxyFetchResponse> =>
+      ipcRenderer.invoke(IpcChannels.net.fetch, request),
   },
 
   feedback: {
