+++ packages/pi-starling-extension/ui/index.html
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+<html lang="en">
+<head><meta charset="UTF-8" /><title>Sero Starling (remote)</title></head>
+<body><div id="root"></div></body>
+</html>
