+++ packages/pi-starling-extension/ui/lib/crypto.ts
@@ -1,0 +1,65 @@
+/**
+ * AES-GCM token encryption with PIN-derived key (PBKDF2).
+ */
+
+async function deriveKey(pin: string, salt: Uint8Array): Promise<CryptoKey> {
+  const enc = new TextEncoder();
+  const keyMaterial = await crypto.subtle.importKey(
+    'raw',
+    enc.encode(pin),
+    'PBKDF2',
+    false,
+    ['deriveKey'],
+  );
+  return crypto.subtle.deriveKey(
+    { name: 'PBKDF2', salt: salt.buffer as ArrayBuffer, iterations: 100000, hash: 'SHA-256' },
+    keyMaterial,
+    { name: 'AES-GCM', length: 256 },
+    false,
+    ['encrypt', 'decrypt'],
+  );
+}
+
+export async function encryptToken(
+  token: string,
+  pin: string,
+): Promise<{ encrypted: string; salt: string; iv: string }> {
+  const salt = crypto.getRandomValues(new Uint8Array(16));
+  const iv = crypto.getRandomValues(new Uint8Array(12));
+  const key = await deriveKey(pin, salt);
+  const ciphertext = await crypto.subtle.encrypt(
+    { name: 'AES-GCM', iv },
+    key,
+    new TextEncoder().encode(token),
+  );
+  return {
+    encrypted: btoa(String.fromCharCode(...new Uint8Array(ciphertext))),
+    salt: btoa(String.fromCharCode(...salt)),
+    iv: btoa(String.fromCharCode(...iv)),
+  };
+}
+
+export async function decryptToken(
+  encrypted: string,
+  salt: string,
+  iv: string,
+  pin: string,
+): Promise<string> {
+  const saltBytes = Uint8Array.from(atob(salt), (c) => c.charCodeAt(0));
+  const ivBytes = Uint8Array.from(atob(iv), (c) => c.charCodeAt(0));
+  const cipherBytes = Uint8Array.from(atob(encrypted), (c) => c.charCodeAt(0));
+  const key = await deriveKey(pin, saltBytes);
+  const plaintext = await crypto.subtle.decrypt(
+    { name: 'AES-GCM', iv: ivBytes },
+    key,
+    cipherBytes,
+  );
+  return new TextDecoder().decode(plaintext);
+}
+
+export async function hashPin(pin: string): Promise<string> {
+  const hash = await crypto.subtle.digest('SHA-256', new TextEncoder().encode(pin));
+  return Array.from(new Uint8Array(hash))
+    .map((b) => b.toString(16).padStart(2, '0'))
+    .join('');
+}
